+++ src/frontend/Player/createPlayer.spec.ts
@@ -1,3 +1,4 @@
+import { Video } from '../types/tracks';
 import { createPlayer } from './createPlayer';
 import { createPlyrPlayer } from './createPlyrPlayer';
 
@@ -20,12 +21,14 @@ describe('createPlayer', () => {
   it('creates a plyr instance when type player is plyr', () => {
     const ref = 'ref' as any;
     const dispatchPlayerTimeUpdate = jest.fn();
+    const video = {} as Video;
 
-    createPlayer('plyr', ref, dispatchPlayerTimeUpdate);
+    createPlayer('plyr', ref, dispatchPlayerTimeUpdate, video);
 
     expect(createPlyrPlayer).toHaveBeenCalledWith(
       ref,
       dispatchPlayerTimeUpdate,
+      video,
     );
   });
 });