+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+- Start ABR at a lower level corresponding to 480p
+
 ## [3.1.6] - 2019-10-24
 
 ### Fixed