+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [3.1.7] - 2019-11-05
+
 ### Fixed
 
 - Start ABR at a lower level corresponding to 480p
@@ -429,7 +431,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Minor fixes and improvements on features and tests
 
-[unreleased]: https://github.com/openfun/marsha/compare/v3.1.6...master
+[unreleased]: https://github.com/openfun/marsha/compare/v3.1.7...master
+[3.1.7]: https://github.com/openfun/marsha/compare/v3.1.6...v3.1.7
 [3.1.6]: https://github.com/openfun/marsha/compare/v3.1.5...v3.1.6
 [3.1.5]: https://github.com/openfun/marsha/compare/v3.1.4...v3.1.5
 [3.1.4]: https://github.com/openfun/marsha/compare/v3.1.3...v3.1.4